requirements.txt
fastapi==0.104.1
uvicorn[standard]
sqlalchemy
pyrogram
python-multipart
websockets
python-dotenv==1.0.0
sqlalchemy==2.0.23
pyrogram==2.0.106
python-dateutil==2.8.2
pydantic==2.4.2
alembic==1.12.1
